requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap5==2023.10
Django==5.0
django-bootstrap5==23.3
django-cleanup==8.0.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
psycopg==3.1.16
psycopg-binary==3.1.16
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
